+++ client/src/pages/app/mainApp.wrapper.tsx
@@ -9,11 +9,13 @@ function MainAppWrapper() {
   useEffect(() => {
     if(!isAuthenticated())
       redirect("/login")
+
+    document.body.style.backgroundColor = "#f4f4ff"
   }, [])
   
   return (
       <Routes>
-        <Route path="/" element={<Navigate to="/" replace />}></Route>
+        <Route path="/" element={<Navigate to="/browse" replace />}></Route>
         <Route path="/browse" element={<BrowseCarsPage />}></Route>
         <Route path="/add" element={<AddCarForm />}></Route>
         <Route path="*" element={<NotFoundPage/>}></Route>
